Create index.html
This is duel between two samurai, that I made using photoshop in shapes. I was inspired by action movie and shows using weapons like swords, especially their interesting camera angles which is what inspired the the panels in my comic. And the main troubles in this was getting some of the more difficult shapes right.